custom-elements/custom-elements-registry/get.html

<!DOCTYPE html>
<title>Custom Elements: CustomElementsRegistry.get</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<iframe id="iframe"></iframe>
<script>
'use strict';
(() => {
  // https://html.spec.whatwg.org/multipage/scripting.html#dom-customelementsregistry-whendefined

  // Use window from iframe to isolate the test.
  const testWindow = iframe.contentDocument.defaultView;
  const customElements = testWindow.customElements;

  let testable = false;
  test(() =>  {
    assert_true('customElements' in testWindow, '"window.customElements" exists');
    assert_true('get' in customElements, '"window.customElements.get" exists');
    testable = true;
  }, '"window.customElements.get" should exist');
  if (!testable)
    return;

  // 1. If this CustomElementsRegistry contains an entry with name name,
  // then return that entry's constructor.
  test(() => {
    const name = 'test-get-existing';
    class C {};
    customElements.define(name, C);
    assert_equals(customElements.get(name), C);
  }, 'get() returns the constructor');

  // 2. Otherwise, return undefined.
  test(() => {
    assert_equals(customElements.get('test-get-not-defined'), undefined);
  }, 'get() returns undefined if the name is not defined');
})();
</script>
</body>
